Fixed #187, The server is not crashing now.
Fixed incorrect .jar path, again...
