+++ src/main/java/mrjake/aunis/event/AunisEventHandler.java
@@ -4,8 +4,6 @@ import mrjake.aunis.Aunis;
 import mrjake.aunis.block.AunisBlocks;
 import mrjake.aunis.raycaster.RaycasterDHD;
 import net.minecraft.block.Block;
-import net.minecraft.client.gui.GuiButton;
-import net.minecraft.client.gui.GuiScreenOptionsSounds;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.ResourceLocation;
@@ -16,14 +14,12 @@ import net.minecraft.world.storage.loot.LootEntryTable;
 import net.minecraft.world.storage.loot.LootPool;
 import net.minecraft.world.storage.loot.RandomValueRange;
 import net.minecraft.world.storage.loot.conditions.LootCondition;
-import net.minecraftforge.client.event.GuiScreenEvent.InitGuiEvent;
 import net.minecraftforge.event.LootTableLoadEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent.RightClickBlock;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent.RightClickEmpty;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent.RightClickItem;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
-import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
 @EventBusSubscriber
@@ -88,14 +84,5 @@ public class AunisEventHandler {
 	    	
 	    	event.getTable().addPool(pool);
 	    }
-	}
-	
-	@SubscribeEvent(priority = EventPriority.LOW)
-	public static void onInitGui(InitGuiEvent event) {
-		if (event.getGui() instanceof GuiScreenOptionsSounds && event.getButtonList().size() > 13) {
-			for (GuiButton button : event.getButtonList().subList(11, event.getButtonList().size())) {
-				button.y += 24;
-			}
-		}
 	}
 }
