+++ src/main/java/mrjake/aunis/event/AunisEventHandlerClient.java
@@ -5,12 +5,16 @@ import mrjake.aunis.AunisProps;
 import mrjake.aunis.block.AunisBlocks;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.client.gui.GuiButton;
+import net.minecraft.client.gui.GuiScreenOptionsSounds;
 import net.minecraft.util.math.RayTraceResult;
 import net.minecraftforge.client.event.DrawBlockHighlightEvent;
+import net.minecraftforge.client.event.GuiScreenEvent.InitGuiEvent;
 import net.minecraftforge.common.config.Config.Type;
 import net.minecraftforge.common.config.ConfigManager;
 import net.minecraftforge.fml.client.event.ConfigChangedEvent.OnConfigChangedEvent;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
+import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.relauncher.Side;
 
@@ -42,4 +46,13 @@ public class AunisEventHandlerClient {
 			event.setCanceled(cancelled);
 		}
     }
+	
+	@SubscribeEvent(priority = EventPriority.LOW)
+	public static void onInitGui(InitGuiEvent event) {
+		if (event.getGui() instanceof GuiScreenOptionsSounds && event.getButtonList().size() > 13) {
+			for (GuiButton button : event.getButtonList().subList(11, event.getButtonList().size())) {
+				button.y += 24;
+			}
+		}
+	}
 }
